requirements.txt
tabulate
numpy
tensorflow-gpu==1.8.0
h5py==2.8.0
Pillow
matplotlib
keras